Add injury analysis feature, onboarding flow, and agent patch capability
- Add full injury analysis pipeline: body map selection, pain assessment,
  injury analyzer engine, analysis results, and rehab plan generation
- Add 6-step onboarding flow with profile management (basic info, medical
  history, surgical history, injury history, activity level, review)
- Add recovery notes feature with Firestore persistence
- Fix build errors across 12+ files from agent-generated code conflicts
- Add surgical edit (patch) capability to agent: find-and-replace with
  3-tier matching (exact, whitespace-normalized, fuzzy) so the agent can
  fix pre-existing files without rewriting them from scratch
- Update orchestrator and iOS dev prompts with patch schema and rules
- Add design review prompt and build retry improvements to agent pipeline

+++ ios/PT-Helper/PT-Helper/Models/WorkoutSession.swift
@@ -4,6 +4,6 @@ struct WorkoutSession {
     var id: UUID
     var date: Date
     var duration: TimeInterval
-    var painLevel: Int
+    var painLevel: Double
     var isCompleted: Bool
 }
